feat: modify ArtContent structure
