+++ src/components/main/ArtContent/ArtContent.tsx
@@ -1,10 +1,12 @@
 import React, { useState } from 'react';
 import { CategoryFilterType, SortFilterType } from './ArtContent.constant';
 import ArtistSection from './ArtistSection/ArtistSection';
-import ArtContentHeader from './ArtContentHeader';
 import ArtSection from '@/components/main/ArtSection';
 import TopRankArtSection from './TopRankArtSection';
-
+import Text from '@/components/common/Text';
+import { COLORS } from '@/constants/styles';
+import SortingMenu from '@/components/common/SortingMenu/SortingMenu';
+import ArtCategoryMenu from '@/components/common/ArtCategoryMenu';
 import * as styles from './ArtContent.style';
 
 export interface FilterType {
@@ -18,9 +20,36 @@ const ArtContent = () => {
     sort: 'recommend',
   });
 
+  const { category: cFilter, sort: sFilter } = filter;
+
+  const onSelectCategoryFilter = (c: CategoryFilterType) => {
+    setFilter((prev) => ({ ...prev, category: c }));
+  };
+
+  const onSelectSortFilter = (s: SortFilterType) => {
+    setFilter((prev) => ({ ...prev, sort: s }));
+  };
+
   return (
     <styles.Wrapper>
-      <ArtContentHeader filter={filter} setFilter={setFilter} />
+      <styles.HeaderWrapper>
+        <styles.Header>
+          <styles.TitleWrap>
+            <Text color={COLORS.main.black} textStyleName="title">
+              취향 발견
+            </Text>
+            <Text color={COLORS.font.black60} textStyleName="subtitle">
+              작가들의 감각적인 전시를 살펴보세요.
+            </Text>
+          </styles.TitleWrap>
+          <ArtCategoryMenu
+            menu={cFilter}
+            onClickMenu={onSelectCategoryFilter}
+          />
+          <SortingMenu menu={sFilter} onClickMenu={onSelectSortFilter} />
+        </styles.Header>
+      </styles.HeaderWrapper>
+
       <TopRankArtSection />
       <ArtistSection />
       <ArtSection />
